python-logic/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
cached-property==1.5.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
email-reply-parser==0.5.12
exchangelib==5.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
lxml==5.2.1
msal==1.28.0
oauthlib==3.2.2
openai==1.24.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyJWT==2.8.0
pyspnego==0.10.2
python-dotenv==1.0.1
requests==2.31.0
requests-ntlm==1.2.0
requests-oauthlib==2.0.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wheel==0.43.0
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
cached-property==1.5.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
email-reply-parser==0.5.12
exchangelib==5.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
lxml==5.2.1
msal==1.28.0
oauthlib==3.2.2
openai==1.24.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyJWT==2.8.0
pyspnego==0.10.2
python-dotenv==1.0.1
requests==2.31.0
requests-ntlm==1.2.0
requests-oauthlib==2.0.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wheel==0.43.0
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
cached-property==1.5.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
email-reply-parser==0.5.12
exchangelib==5.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
lxml==5.2.1
msal==1.28.0
oauthlib==3.2.2
openai==1.30.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyJWT==2.8.0
pyspnego==0.10.2
python-dotenv==1.0.1
requests==2.31.0
requests-ntlm==1.2.0
requests-oauthlib==2.0.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wheel==0.43.0
